requirements.txt
opencv-python>=4.8
numpy>=1.24
pyserial>=3.5